requirements.txt
Django==1.8.4
djangorestframework==3.3.0
django-extensions==1.6.1
djangorestframework-jwt==1.8.0
mysqlclient==1.3.9
serpy==0.1.0
xlwt==1.1.2
xlrd==0.9.4
pytz==2016.2
pycrypto==2.6.1
python-dateutil==2.5.3
PyJWT==1.4.0
celery==3.1.19
django-bulk-update==1.1.10
redis==2.10.3
requests==2.10.0
pymongo==3.3.0
raven==5.27.1
elasticsearch==2.4.0
uWSGI==2.0.13.1
XlsxWriter==0.9.3
ipcalc==1.99.0
beautifulsoup4==4.5.3
